Added function and imported path module (see description)
The read stream used is asynchronous.
Problem: The data needed wouldn't be fully loaded when the application starts.
Solution: Created `loadPlanetsData` function; it returns a promise, which consists of the parsing logic. It would be added to `server.js` as an awaited result right before the server starts accepting requests.

The path module was used to create the dataset's relative path.