Applied separation of concerns
For launches.model.js, a getter function was defined to abstract the implementation of retrieving the data since the model focuses on the logic related to the data. This function is exported instead of the data itself.
launches.controller.js makes use of the launches model, so the function from the model is imported. The function defined in the controller has been renamed to "httpGetAllLaunches" because this function works directly with the request and returns the response to the client.
Since the router is related to the controller, the renamed function from controller has been imported to launches.router.js as needed.